requirements.txt
azure-search-documents
openai
python-dotenv
numpy
azure-core
PyPDF2
tiktoken
python-docx
pandas
python-pptx
openpyxl
azure-storage-blob
azure-identity